# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user_handler.UpdateAvatarRequest:
    properties:
      avatar_url:
        type: string
    required:
    - avatar_url
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Gateway for Pinstack social media platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pinstack API Gateway
  version: "1.0"
paths:
...
  /users/{id}/avatar:
    put:
      consumes:
      - application/json
      description: Update the avatar URL for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Avatar update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user_handler.UpdateAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Avatar updated successfully
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Operation not allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user avatar
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"